requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
stix2>=3.0.1
taxii2-client>=2.3.0
